Use stdout for logging output
Use stdout instead of stderr for outputting messages. While using stderr
is default e.g. in Python as well, using it for add-ons makes the logs
in Home Assistant OS (which uses Docker with journald log driver) to be
logged with err priority (level 3) to the Systemd Journal.

Moreover, as a result of using different pipe than stdout, mixing
standard output (e.g. echo commands) with bashio logging might result in
the add-on output being printed out of order once the pipes are printed
next to each other, unless the output has been printed to stderr as
well. This can cause confusion, as seen in #154.